Create Tester for volume
Testing the volume reading. Has two different AudioContexts of the same file one with Gain level 4 the other with Gain level 1 then ran both and divided the input of the first by the second to get a ratio between the two volumes. The ratio was 4.0 which is exactly what was desired. Can now move on to volume control given current reading